Add true AutoQuad flight mode status display to toolbar and voice announcements (AH/PH/DVH, etc); Add RSSI status to main toobar; Give toolbar widgets a disabled style when no MAV connection; Add delay before GPS lock lost/established voice notices to avoid announcing momentary lapses. NOTE: Must run `git submodule update` before building to pull in latest AQ MAVLink definitions.

+++ src/ui/uas/UASControlWidget.h
@@ -62,7 +62,7 @@ public slots:
     /** @brief Update the mode */
     void updateMode(int uas,QString mode,QString description);
     /** @brief Update state */
-    void updateState(int state);
+    void updateState(uint8_t state);
     /** @brief Update internal state machine */
     void updateStatemachine();
